+++ package.json
@@ -9,6 +9,7 @@
     "react-scripts": "4.0.3",
     "react-slick": "^0.28.1",
     "slick-carousel": "^1.8.1",
+    "swiper": "^7.3.3",
     "typescript": "^4.5.2"
   },
   "scripts": {
