+++ public/index.html
@@ -8,11 +8,11 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta
-      name="Strapi and ReactJs Boilerplate"
+      name="CCB Social"
       content="A Strapi and ReactJs Boilerplate for CITi Onboarding projects."
     />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>Strapi and ReactJs Boilerplate</title>
+    <title>CCB Social</title>
   </head>
   <body>
     <div id="root"></div>
