+++ src/App.jsx
@@ -1,13 +1,7 @@
 import React from 'react';
 import './style.css';
-import Banner from './Views/Banner/Banner';
-import NossosParceiros from './Views/Nossos Parceiros/NossosParceiros';
-import Carousel from './Views/Projetos/carousel';
-import Campanha from './Views/Campanha';
-import Donate from './Views/doacao/doacao';
-import NossaMissao from './Views/NossaMissao/index'
-import NavBar from './Views/NavBar';
-import Rodape from './Views/Rodape';
+import {Banner, NavBar, NossaMissao, Projetos, Campanha, Rodape} from './Views';
+
 
 const App = () => {
   return (
@@ -15,11 +9,9 @@ const App = () => {
       <NavBar/>
       <Banner/>
       <NossaMissao/>
-      <Carousel/>
+      <Projetos/>
       <Campanha/>
-      <Donate/>
       <Rodape/>
-
   </div>
 )
   }
