--- src/Views/Banner/index.js
@@ -1,2 +1,0 @@
-import { Banner } from "./Banner"
-export default Banner
