--- src/Views/Banner/style.css
@@ -1,109 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
-
-* {
-        margin: 0px;
-    }
-    
-    /*Imagem e fundo do Banner*/
-
-    .containerBanner{
-        width: 100%;
-        height: 100vh;
-        text-align: left;
-        justify-content:left;
-        align-items: left;
-        display: flex;
-    }
-    
-    .cardBanner{
-        background: linear-gradient(90deg, rgba(2, 116, 225, 0.8) 10%, rgba(255, 255, 255, 0) 100%);;  
-    }
-
-    .imageBanner{
-        opacity: 50%;
-        position:absolute;
-        top: 0;
-        left: 0;
-        height: 100vh;
-        width: 100%;
-        z-index: 0;
-    } 
-    
-    /*Título e subtítulo do Banner*/
-    
-    .infosBanner {
-        display: flex;
-        align-items: flex-start;
-        
-    }
-
-    .tituloBanner {
-        font-family: 'Montserrat';
-        font-style: normal;
-        font-weight: bold;
-        font-size: 40px;
-        line-height: 49px;
-        text-align: left;
-        color: #FFFFFF;
-        text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-
-        padding-left: 108px;
-        
-        width: 498px;
-        height: 98px;
-        left: 108px;
-        top: 294px;
-
-        z-index: 1;
-    }
-
-    .subtituloBanner {
-        font-family: 'Montserrat';
-        font-style: normal;
-        font-weight: bold;
-        text-align: left;
-        font-size: 16px;
-        line-height: 20px;
-        color: #FFFFFF;
-        text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-        
-        padding-left: 108px;
-
-        width: 377px;
-        height: 40px;
-        left: 108px;
-        top: 422px;
-        
-        z-index: 1;
-    }
-
-    /*Botão do Banner*/
-
-    .setaBanner{
-        margin-left: 10px;
-        
-    }
-
-    .linkBanner{
-        text-decoration: none;
-        color: #FFFFFF;
-        
-    }
-
-    .doacaoBanner{
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        background: #FDB600;
-        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-        border-radius: 10px;
-        width: 152px;
-        height: 62px;
-        margin-top: 25px;
-        z-index: 1;
-        font-size: 20px;
-        
-        
-    }
-    
-
