--- src/Views/Campanha/index.js
@@ -1,2 +1,0 @@
-import { Campanha } from "./Campanha"
-export default Campanha;
