+++ src/Views/Campanha/index.jsx
@@ -35,4 +35,3 @@ export const Campanha = () => {
       
     )
 }
-export default Campanha;
