+++ src/Views/Campanha/style.css
@@ -1,27 +1,16 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
-
-* {
-    margin: 0px;
-}
-
 .containerCampanha {  
     height: 100vh;
     background-color:#F4F4F4;
-    padding-top: 117px;
-    
     display: flex;
     flex-direction: row-reverse;
     align-items: center;
     justify-content: space-between;
-    
 }
 
 .infosCampanha {
-    
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    
 }
 
 .tituloCampanha {
