--- src/Views/NavBar/index.js
@@ -1,2 +1,0 @@
-import { NavBar } from "./NavBar";
-export default NavBar
