+++ src/Views/NavBar/style.css
@@ -1,60 +1,56 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat&family=Roboto:wght@100&display=swap');
-.spacing{
-    padding: 0;
+.NavbarContainer {
     display: flex;
     justify-content: center;
-    text-align: center;
+    position: fixed;
+    height: 100px;
+    width: 100%;
+    background-color: white;
+    z-index: 100;
+    padding: 0 40px 0;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);;
+}
+
+.spacing{
+    display: flex;
     color: black;
     align-items: center;
 }
-.Navbar {
-    display: flex;
-    flex-direction: row;
-    position: fixed;
-    top: 0px;
-    height: 75px;
-    background-color: white;
-    z-index: 100;
 
-}
-.Navbar a{
+.NavbarContainer a{
     margin-left: 30px;
     text-decoration: none;
-    font-family: Montserrat;
-    font-style: normal;
-    font-weight: normal;
     font-size: 16px;
-    line-height: 20px;
-    text-align: center;
     color: #000000;
+    transition: 0.5s;
+    border-bottom: 2px solid white;
 }
 
+.NavbarContainer a:hover {
+    color: #0849D8;
+    border-bottom: 2px solid #0849D8;
+}
 
 .doar {
-    font-family: Montserrat;
-    font-style: normal;
-    font-weight: bold;
     font-size: 20px;
-    line-height: 24px;
-    text-align: right;
     color: #FFFFFF;
 }
 
-.button_navbar {
-    width: 111px;
-    height: 54px;
-    left: 1225px;
-    top: 31px;
+.ButtonNavbar {
+    border: none;
+    cursor: pointer;
+    width: 110px;
+    height: 55px;
     background: #0849D8;
     border-radius: 10px;
-    font-family: Montserrat;
-    font-style: normal;
     font-weight: bold;
     font-size: 20px;
-    line-height: 24px;
     color: #FFFFFF;
+    transition: 0.5s;
 }
 
+.ButtonNavbar:hover {
+    background: #082c7a;
+}
