--- src/Views/NossaMissao/index.js
@@ -1,2 +1,0 @@
-import  {NossaMissao}  from "./NossaMissao"
-export default NossaMissao
