+++ src/Views/Projetos/style.css
@@ -1,34 +1,25 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat&display=swap');
 
-.divCarousel{
-  max-width: 85%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  justify-content: center;
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
+.containerProjetos{
+  width: 100%;
   height: 100vh;
-  margin-top: 200px;
+  background-color: #FFFFFF;
 }
 
-.cardC .slick-slide{
-  margin: 10px ;
+.carouselProjetos {
+  margin-left: 170px;
 }
-#footer{
-  height: 400px;
+.carouselProjetos .slick-slider {
+  width: 90%;
+  padding-bottom: 10px;
+  height: 500px;
 }
-
-.slick-prev {
-  display: flex;
-}
-.slick-next {
-  
+.carouselProjetos .slick-list {
+  padding-bottom: 10px;
 }
 .slick-prev:before {
   content: url('../../assets/right-arrow.svg')!important; 
   content: url('../../assets/inside-left-arrow.svg')!important;
-  
 }
 
 .slick-next:before {
@@ -37,45 +28,40 @@
 }
 
 .cardC{
-  
   background-color: #FFFFFF;
   border-radius: 15px;
-  box-shadow: rgba(0, 0, 0, 0.15) 5.95px 5.95px 2.6px;
-  border: white 2.5px solid;
-  width: 300px !important;
-  height: 500px !important;
+  box-shadow: 4px 4px 4px rgba(0, 0, 0, 0.25);
+  width: 320px !important;
+  height: 480px !important;
+}
 
-}
 .cardC img{
-  border-radius: 10px;
+  border-radius: 15px 15px 0 0;
 }
+
 .imgC{
-  width: 300px;
+  width: 320px;
   max-width: fit-content;
   height: 200px;
-  min-width: 300px;
+  min-width: 320px;
 }
 .pC {
   font-family: Montserrat;
   font-style: normal;
   font-weight: normal;
-  font-size: 16px;
+  font-size: 14px;
   line-height: 20px;
   text-align: center;
-  margin-top: 24px;
+  margin: 24px 20px 0;
   color: #000000;
 }
-#h1c{
-  padding-bottom: 100px;
-  font-family: Montserrat;
-  font-style: normal;
+
+.titleProjetos {
   font-weight: 600;
   font-size: 36px;
-  line-height: 44px;
   text-align: center;
   color: #000000;
-  margin-bottom: -100px;
- 
+  padding: 150px 0 50px; 
 }
 .h1C {
 font-family: Montserrat;
