--- src/Views/Rodape/index.js
@@ -1,2 +1,0 @@
-import { Rodape } from "./Rodape";
-export default Rodape
