+++ src/Views/Rodape/index.jsx
@@ -1,4 +1,4 @@
-import react from "react";
+import React from "react";
 import "./style.css";
 
 export const Rodape = () => {
