+++ src/Views/index.ts
@@ -1,0 +1,8 @@
+import { NavBar } from './NavBar';
+import { Banner } from './Banner';
+import { NossaMissao } from './NossaMissao'
+import { Projetos } from './Projetos';
+import { Campanha } from './Campanha';
+import { Rodape } from './Rodape';
+
+export {Banner, NavBar, NossaMissao, Projetos, Campanha, Rodape}
